requirements.txt
streamlit
pandas
python-binance
streamlit-autorefresh
